+++ homepage/homepage/components/nav.tsx
@@ -2,7 +2,6 @@ import { ThemeToggle } from "@/components/ThemeToggle";
 import { socials } from "@/lib/socials";
 import { JazzLogo } from "gcmp-design-system/src/app/components/atoms/logos/JazzLogo";
 import { Nav } from "gcmp-design-system/src/app/components/organisms/Nav";
-import { BookTextIcon, BoxIcon, CodeIcon } from "lucide-react";
 import { DocNav } from "./docs/nav";
 
 export function JazzNav() {
@@ -13,46 +12,17 @@ export function JazzNav() {
       items={[
         { title: "Jazz Cloud", href: "/cloud" },
         {
-          title: "Documentation",
+          title: "Docs",
           href: "/docs",
-          items: [
-            {
-              icon: (
-                <BookTextIcon
-                  className="size-5 stroke-blue dark:stroke-blue-500 shrink-0"
-                  strokeWidth={1.5}
-                />
-              ),
-              title: "Documentation",
-              href: "/docs",
-              description:
-                "Get started with using Jazz by learning the core concepts, and going through guides.",
-            },
-            {
-              icon: (
-                <CodeIcon
-                  className="size-5 stroke-blue dark:stroke-blue-500 shrink-0"
-                  strokeWidth={1.5}
-                />
-              ),
-              title: "Example apps",
-              href: "/examples",
-              description:
-                "Demo and source code for example apps built with Jazz.",
-            },
-            {
-              icon: (
-                <BoxIcon
-                  className="size-5 stroke-blue dark:stroke-blue-500 shrink-0"
-                  strokeWidth={1.5}
-                />
-              ),
-              title: "API reference",
-              href: "/api-reference",
-              description:
-                "API references for packages like jazz-tools, jazz-react, and more.",
-            },
-          ],
+          items: [],
+        },
+        {
+          title: "Examples",
+          href: "/examples",
+        },
+        {
+          title: "API ref",
+          href: "/api-reference",
         },
         {
           title: "Built with Jazz",