+++ packages/cojson-storage-indexeddb/src/tests/idbNode.test.ts
@@ -15,11 +15,7 @@ test("Should be able to initialize and load from empty DB", async () => {
     Crypto,
   );
 
-  node.addPeer(await IDBStorage.asPeer({ trace: true }));
-
-  console.log("yay!");
-
-  const _group = node.createGroup();
+  await node.addPeer(await IDBStorage.asPeer({}));
 
   await new Promise((resolve) => setTimeout(resolve, 200));
 
@@ -35,11 +31,9 @@ test("Should be able to sync data to database and then load that from a new node
     Crypto,
   );
 
-  node1.addPeer(
+  await node1.addPeer(
     await IDBStorage.asPeer({ trace: true, localNodeName: "node1" }),
   );
-
-  console.log("yay!");
 
   const group = node1.createGroup();
 
@@ -55,9 +49,7 @@ test("Should be able to sync data to database and then load that from a new node
     Crypto,
   );
 
-  node2.addPeer(
-    await IDBStorage.asPeer({ trace: true, localNodeName: "node2" }),
-  );
+  await node2.addPeer(await IDBStorage.asPeer({ localNodeName: "node2" }));
 
   const map2 = await node2.load(map.id);
   if (map2 === "unavailable") {