+++ packages/jazz-browser/src/index.ts
@@ -204,7 +204,7 @@ export function provideBrowserLockSession(
 /** @category Invite Links */
 export function createInviteLink<C extends CoValue>(
   value: C,
-  role: "reader" | "writer" | "admin",
+  role: "reader" | "writer" | "admin" | "writeOnly",
   // default to same address as window.location, but without hash
   {
     baseURL = window.location.href.replace(/#.*$/, ""),
@@ -218,7 +218,9 @@ export function createInviteLink<C extends CoValue>(
     currentCoValue = currentCoValue.getGroup().core;
   }
 
-  if (currentCoValue.header.ruleset.type !== "group") {
+  const { ruleset, meta } = currentCoValue.header;
+
+  if (ruleset.type !== "group" || meta?.type === "account") {
     throw new Error("Can't create invite link for object without group");
   }