+++ packages/jazz-run/src/createWorkerAccount.ts
@@ -1,4 +1,3 @@
-import { emptyKnownState } from "cojson";
 import { createWebSocketPeer } from "cojson-transport-ws";
 import { Account, WasmCrypto, isControlledAccount } from "jazz-tools";
 import { WebSocket } from "ws";
@@ -28,22 +27,10 @@ export const createWorkerAccount = async ({
     throw new Error("account is not a controlled account");
   }
 
-  const accountCoValue = account._raw.core;
-  const accountProfileCoValue = account.profile!._raw.core;
-  const syncManager = account._raw.core.node.syncManager;
-
-  await Promise.all([
-    syncManager.syncCoValue(accountCoValue, emptyKnownState(accountCoValue.id)),
-    syncManager.syncCoValue(
-      accountProfileCoValue,
-      emptyKnownState(accountCoValue.id),
-    ),
-  ]);
-
   await account.waitForAllCoValuesSync({ timeout: 4_000 });
 
   return {
-    accountId: account.id,
+    accountID: account.id,
     agentSecret: account._raw.agentSecret,
   };
 };