+++ packages/jazz-browser/CHANGELOG.md
@@ -1,5 +1,65 @@
 # jazz-browser
 
+## 0.8.49
+
+### Patch Changes
+
+- Updated dependencies [25dfd90]
+  - cojson@0.8.49
+  - cojson-storage-indexeddb@0.8.49
+  - cojson-transport-ws@0.8.49
+  - jazz-tools@0.8.49
+
+## 0.8.48
+
+### Patch Changes
+
+- Updated dependencies [635e824]
+- Updated dependencies [10ea733]
+- Updated dependencies [0a85982]
+  - jazz-tools@0.8.48
+  - cojson@0.8.48
+  - cojson-storage-indexeddb@0.8.48
+  - cojson-transport-ws@0.8.48
+
+## 0.8.45
+
+### Patch Changes
+
+- 6f0bd7f: Throw an error if an invite is created from an account owned coValue
+- Updated dependencies [6f0bd7f]
+- Updated dependencies [fca6a0b]
+- Updated dependencies [fa41f8e]
+- Updated dependencies [88d7d9a]
+- Updated dependencies [60e35ea]
+  - cojson@0.8.45
+  - jazz-tools@0.8.45
+  - cojson-storage-indexeddb@0.8.45
+  - cojson-transport-ws@0.8.45
+
+## 0.8.44
+
+### Patch Changes
+
+- Updated dependencies [5d20c81]
+  - cojson@0.8.44
+  - cojson-storage-indexeddb@0.8.44
+  - cojson-transport-ws@0.8.44
+  - jazz-tools@0.8.44
+
+## 0.8.41
+
+### Patch Changes
+
+- ac216b9: Add a new writeOnly role, to limit access only to their own changes. Useful to push objects into lists of moderated content.
+- Updated dependencies [3252502]
+- Updated dependencies [6370348]
+- Updated dependencies [ac216b9]
+  - cojson@0.8.41
+  - cojson-storage-indexeddb@0.8.41
+  - cojson-transport-ws@0.8.41
+  - jazz-tools@0.8.41
+
 ## 0.8.40
 
 ### Patch Changes