+++ packages/jazz-react-native-media-images/CHANGELOG.md
@@ -1,5 +1,40 @@
 # jazz-browser-media-images
 
+## 0.8.49
+
+### Patch Changes
+
+- jazz-tools@0.8.49
+
+## 0.8.48
+
+### Patch Changes
+
+- Updated dependencies [635e824]
+- Updated dependencies [0a85982]
+  - jazz-tools@0.8.48
+
+## 0.8.45
+
+### Patch Changes
+
+- Updated dependencies [fa41f8e]
+- Updated dependencies [88d7d9a]
+- Updated dependencies [60e35ea]
+  - jazz-tools@0.8.45
+
+## 0.8.44
+
+### Patch Changes
+
+- jazz-tools@0.8.44
+
+## 0.8.41
+
+### Patch Changes
+
+- jazz-tools@0.8.41
+
 ## 0.8.39
 
 ### Patch Changes